+++ src/screens/FeedScreen/hooks/useCoreFeed.js
@@ -4,21 +4,15 @@ import axios from 'axios';
 import {useSafeAreaInsets} from 'react-native-safe-area-context';
 
 import StorageUtils from '../../../utils/storage';
-import dimen from '../../../utils/dimen';
+import useViewPostTimeHook from './useViewPostTimeHook';
 import {Context} from '../../../context';
 import {FEEDS_CACHE} from '../../../utils/cache/constant';
-import {SOURCE_FEED_TAB} from '../../../utils/constants';
 import {checkIsHasColor, hexToRgb} from '../../../utils/colors';
 import {downVote, upVote} from '../../../service/vote';
-import {getFeedDetail, getMainFeedV2WithTargetFeed, viewTimePost} from '../../../service/post';
+import {getFeedDetail, getMainFeedV2WithTargetFeed} from '../../../service/post';
 import {listFeedColor} from '../../../configs/FeedColor';
 import {saveToCache} from '../../../utils/cache';
-import {
-  setFeedByIndex,
-  setMainFeeds,
-  setTimer,
-  setViewPostTimeIndex
-} from '../../../context/actions/feeds';
+import {setFeedByIndex, setMainFeeds, setTimer} from '../../../context/actions/feeds';
 
 const useCoreFeed = () => {
   const [loading, setLoading] = React.useState(false);
@@ -34,6 +28,12 @@ const useCoreFeed = () => {
   const {feeds, timer, viewPostTimeIndex} = feedsContext;
   const {myProfile} = profileContext;
   const {bottom} = useSafeAreaInsets();
+
+  const {sendViewPostTimeWithFeeds, updateViewPostTime, isSamePostViewed} = useViewPostTimeHook(
+    dispatch,
+    timer,
+    viewPostTimeIndex
+  );
 
   const getDataFeeds = async (offsetFeed = 0, useLoading = false, targetFeed = null) => {
     setCountStack(null);
@@ -222,28 +222,7 @@ const useCoreFeed = () => {
   };
 
   const sendViewPostTime = async (withResetTime = false) => {
-    const currentTime = new Date();
-    const diffTime = currentTime.getTime() - timer.getTime();
-    const id = feeds?.[viewPostTimeIndex]?.id;
-    if (!id) return;
-
-    viewTimePost(id, diffTime, SOURCE_FEED_TAB);
-    if (withResetTime) setTimer(new Date(), dispatch);
-  };
-
-  const getCurrentPostViewed = (momentumEvent) => {
-    const {y} = momentumEvent.nativeEvent.contentOffset;
-    const shownIndex = Math.ceil(y / dimen.size.FEED_CURRENT_ITEM_HEIGHT);
-    return shownIndex;
-  };
-
-  const updateViewPostTime = (momentumEvent) => {
-    setViewPostTimeIndex(getCurrentPostViewed(momentumEvent), dispatch);
-    setTimer(new Date(), dispatch);
-  };
-
-  const isSamePostViewed = (momentumEvent) => {
-    return getCurrentPostViewed(momentumEvent) === viewPostTimeIndex;
+    sendViewPostTimeWithFeeds(feeds, withResetTime);
   };
 
   return {
@@ -272,6 +251,7 @@ const useCoreFeed = () => {
     onDeleteBlockedPostCompleted,
     saveSearchHeight,
     sendViewPostTime,
+    sendViewPostTimeWithFeeds,
     setDownVote,
     setIsLastPage,
     setMainFeeds,
